a bit more norm friendy. Now passing all type u. Next up hash-sign flags for types o, x, X

+++ libft/ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: malasalm <malasalm@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/24 16:39:48 by malasalm          #+#    #+#             */
-/*   Updated: 2019/11/11 15:11:15 by malasalm         ###   ########.fr       */
+/*   Updated: 2020/08/11 12:51:18 by malasalm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_intlen(long num)
+int	ft_intlen(long long num)
 {
 	int len;
